refactor(helpertext): move HelperText to its folder with theme interface
#443 #234
