+++ src/lib/components/HelperText/HelperText.tsx
@@ -3,11 +3,9 @@ import type { ComponentProps, FC, PropsWithChildren } from 'react';
 import type { FlowbiteColors } from '../Flowbite/FlowbiteTheme';
 import { useTheme } from '../Flowbite/ThemeContext';
 
-export interface FlowbiteFormControlsTheme {
-  helperText: {
-    base: string;
-    colors: HelperColors;
-  };
+export interface FlowbiteHelperTextTheme {
+  base: string;
+  colors: HelperColors;
 }
 
 export interface HelperColors extends Pick<FlowbiteColors, 'gray' | 'info' | 'failure' | 'warning' | 'success'> {
@@ -19,13 +17,11 @@ export interface HelperTextProps extends PropsWithChildren<Omit<ComponentProps<'
   value?: string;
 }
 
-const HelperText: FC<HelperTextProps> = ({ value, children, color = 'default', className, ...props }) => {
-  const theme = useTheme().theme.formControls.helperText;
+export const HelperText: FC<HelperTextProps> = ({ value, children, color = 'default', className, ...props }) => {
+  const theme = useTheme().theme.helperText;
   return (
     <p className={classNames(theme.base, theme.colors[color], className)} {...props}>
       {value ?? children ?? ''}
     </p>
   );
 };
-
-export default HelperText;
