+++ src/lib/components/index.ts
@@ -6,24 +6,32 @@ export * from './Breadcrumb';
 export * from './Button';
 export * from './Card';
 export * from './Carousel';
+export * from './Checkbox';
 export * from './DarkThemeToggle/DarkThemeToggle';
 export * from './Dropdown';
+export * from './FileInput';
 export * from './Flowbite';
 export type { CustomFlowbiteTheme } from './Flowbite/FlowbiteTheme';
 export * from './Footer';
-export * from './FormControls';
+export * from './HelperText';
+export * from './Label';
 export * from './ListGroup/ListGroup';
 export * from './Modal';
 export * from './Navbar';
 export * from './Pagination';
 export * from './Progress';
+export * from './Radio';
 export * from './Rating';
+export * from './Select';
 export * from './Sidebar/Sidebar';
 export * from './Spinner/Spinner';
 export * from './Tab/Tabs';
 export * from './Table';
+export * from './Textarea';
+export * from './TextInput';
 export * from './Timeline';
 export * from './Toast';
+export * from './ToggleSwitch';
 export * from './Tooltip';
 
 export type DeepPartial<T> = T extends object
