+++ src/lib/theme/default.ts
@@ -539,16 +539,14 @@ const theme: FlowbiteTheme = {
     },
     label: 'ml-3 text-sm font-medium text-gray-900 dark:text-gray-300',
   },
-  formControls: {
-    helperText: {
-      base: 'mt-2 text-sm',
-      colors: {
-        gray: 'text-gray-500 dark:text-gray-400',
-        info: 'text-blue-700 dark:text-blue-800',
-        success: 'text-green-600 dark:text-green-500',
-        failure: 'text-red-600 dark:text-red-500',
-        warning: 'text-yellow-500 dark:text-yellow-600',
-      },
+  helperText: {
+    base: 'mt-2 text-sm',
+    colors: {
+      gray: 'text-gray-500 dark:text-gray-400',
+      info: 'text-blue-700 dark:text-blue-800',
+      success: 'text-green-600 dark:text-green-500',
+      failure: 'text-red-600 dark:text-red-500',
+      warning: 'text-yellow-500 dark:text-yellow-600',
     },
   },
   listGroup: {
